+++ wallet/src/main.rs
@@ -26,6 +26,7 @@ use sp_core::{
 mod amoeba;
 mod money;
 mod sync;
+mod rpc;
 
 /// The default RPC endpoint for the wallet to connect to
 const DEFAULT_ENDPOINT: &str = "http://localhost:9933";
@@ -159,7 +160,7 @@ async fn main() -> anyhow::Result<()> {
     // well as the UTXOs owned by the keys in this wallet.
 
     // Read node's genesis block.
-    let node_genesis = sync::node_get_block_hash(0, &client).await?.expect("node should be able to return some genesis hash");
+    let node_genesis = rpc::node_get_block_hash(0, &client).await?.expect("node should be able to return some genesis hash");
     println!("Node's Genesis block::{:?}", node_genesis);
 
     // Open the database
@@ -193,7 +194,7 @@ async fn main() -> anyhow::Result<()> {
     let mut height:u32 = num_blocks as u32 - 1;
     let hash_ivec = wallet_blocks_tree.get(height.encode())?.expect("block should be present");
     let mut wallet_hash: H256 = H256::decode(&mut &hash_ivec[..])?;
-    let mut node_hash: Option<H256> = sync::node_get_block_hash(height, &client).await?;
+    let mut node_hash: Option<H256> = rpc::node_get_block_hash(height, &client).await?;
 
     println!("About to revert 0 or more blocks. wallet: {wallet_hash:?}. node: {node_hash:?}");
 
@@ -212,7 +213,7 @@ async fn main() -> anyhow::Result<()> {
         // For now we assume the node still knows about these old blocks.
         // Eventually the wallet could consider storing entire blocks locally for
         // the sake of unapplying them.
-        // let _block = sync::node_get_block(wallet_hash, &client).await?.expect("Node should be able to return a block that the wallet previously synced");
+        // let _block = rpc::node_get_block(wallet_hash, &client).await?.expect("Node should be able to return a block that the wallet previously synced");
         // println!("Got block to un_apply");
 
         // Remove the invalid
@@ -224,21 +225,21 @@ async fn main() -> anyhow::Result<()> {
         height -= 1;
         let hash_ivec = wallet_blocks_tree.get(height.encode())?.expect("block should be present");
         wallet_hash = H256::decode(&mut &hash_ivec[..])?;
-        node_hash = sync::node_get_block_hash(height, &client).await?;
+        node_hash = rpc::node_get_block_hash(height, &client).await?;
     }
 
     // Orphaned blocks (if any) have been discarded at this point.
     // So we prepare our variables for forward syncing.
     println!("Resyncing from common ancestor {node_hash:?} - {wallet_hash:?}");
     height += 1;
-    node_hash = sync::node_get_block_hash(height, &client).await?;
+    node_hash = rpc::node_get_block_hash(height, &client).await?;
 
     // Now that we have checked for reorgs and rolled back any orphan blocks, we can go ahead and sync forward.
    while let Some(hash) = node_hash  {
         println!("Forward syncing height {height}, hash {hash:?}");
 
         // Fetch the entire block in order to apply its transactions
-        let block = sync::node_get_block(hash, &client).await?.expect("Node should be able to return a block whose hash it already returned");
+        let block = rpc::node_get_block(hash, &client).await?.expect("Node should be able to return a block whose hash it already returned");
         println!("Got block");
 
         // Apply the new block
@@ -246,7 +247,7 @@ async fn main() -> anyhow::Result<()> {
 
         height += 1;
 
-        node_hash = sync::node_get_block_hash(height, &client).await?;
+        node_hash = rpc::node_get_block_hash(height, &client).await?;
     }
     
     println!("Done with forward sync up to {}", height - 1);
@@ -330,6 +331,7 @@ async fn main() -> anyhow::Result<()> {
     }
 }
 
+//TODO move to rpc.rs
 /// Fetch an output from chain storage given an OutputRef
 pub async fn fetch_storage<V: Verifier>(
     output_ref: &OutputRef,
